Rename to image banner and fix button attention levels
