--- libs/extensions/angular/banner/src/index.ts
@@ -1,1 +1,0 @@
-export { BannerComponent } from './banner.component';
