+++ libs/extensions/angular/image-banner/ng-package.json
