+++ libs/extensions/angular/image-banner/src/banner.component.spec.ts
