+++ libs/extensions/angular/image-banner/src/index.ts
@@ -1,0 +1,1 @@
+export { ImageBannerComponent } from './banner.component';
