+++ tsconfig.base.json
@@ -18,7 +18,7 @@
     "paths": {
       "@kirbydesign/angular/badge": ["libs/angular/dist/badge"],
       "@kirbydesign/core": ["libs/core/dist/"],
-      "@kirbydesign/extensions-angular/banner": ["libs/extensions/angular/dist/banner"],
+      "@kirbydesign/extensions-angular/image-banner": ["libs/extensions/angular/dist/image-banner"],
       "@kirbydesign/designsystem": ["libs/designsystem/dist/"],
       "@kirbydesign/designsystem/*": ["libs/designsystem/dist/*"],
       "@kirbydesign/designsystem/accordion": ["libs/designsystem/dist/accordion"],
